Schema updates - WIP

+++ hbm2ddl-core/src/main/java/com/devskiller/hbm2ddl/SchemaGenerator.java
@@ -11,12 +11,13 @@ import java.nio.file.PathMatcher;
 import java.nio.file.Paths;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.attribute.BasicFileAttributes;
+import java.sql.Connection;
+import java.sql.DriverManager;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.EnumSet;
 import java.util.Enumeration;
 import java.util.List;
-import java.util.Properties;
+import java.util.stream.Collectors;
 
 import org.hibernate.boot.MetadataSources;
 import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
@@ -25,55 +26,73 @@ import org.hibernate.tool.schema.TargetType;
 
 class SchemaGenerator {
 
-	void generate(GenerationMode generationMode, File outputFile, List<String> packages, Action action, Properties jpaProperties, boolean formatOutput, String delimiter) throws Exception {
-		Path schemaPath = outputFile.toPath();
-		Files.deleteIfExists(schemaPath);
-		if (jpaProperties.getProperty("hibernate.dialect") == null) {
-			jpaProperties.setProperty("hibernate.dialect", "org.hibernate.dialect.H2Dialect");
+	private static final String DB_URL = "jdbc:h2:mem:hbm2ddl";
+
+	void generate(GeneratorSettings settings) throws Exception {
+		Path schemaPath = settings.getSchemaFile().toPath();
+		if (settings.getAction() != Action.UPDATE) {
+			Files.deleteIfExists(schemaPath);
+		}
+		if (settings.getJpaProperties().getProperty("hibernate.dialect") == null) {
+			settings.getJpaProperties().setProperty("hibernate.dialect", "org.hibernate.dialect.H2Dialect");
 		}
 
-		if (generationMode == GenerationMode.DATABASE) {
-			jpaProperties.setProperty("hibernate.connection.url", "jdbc:h2:file:/tmp/hbm2ddl");
-			jpaProperties.setProperty("hibernate.connection.username", "sa");
-			jpaProperties.setProperty("hibernate.connection.password", "");
-			jpaProperties.setProperty("javax.persistence.schema-generation.scripts.action", action.toSchemaGenerationAction());
-			jpaProperties.setProperty("javax.persistence.schema-generation.scripts.create-target", outputFile.getAbsolutePath());
-			jpaProperties.setProperty("javax.persistence.schema-generation.scripts.drop-target", outputFile.getAbsolutePath());
+		File outputFile = settings.getSchemaFile();
+		if (settings.getGenerationMode() == GenerationMode.DATABASE) {
+			settings.getJpaProperties().setProperty("hibernate.connection.url", DB_URL);
+			settings.getJpaProperties().setProperty("hibernate.connection.username", "sa");
+			settings.getJpaProperties().setProperty("hibernate.connection.password", "");
+			settings.getJpaProperties().setProperty("javax.persistence.schema-generation.scripts.action", settings.getAction().toSchemaGenerationAction());
 
-// below updates the SCHEMA :)
-// we need to create a database, apply previous script and then run hibernate
-//			jpaProperties.setProperty("javax.persistence.schema-generation.scripts.create-target", outputFile.getAbsolutePath());
-//			jpaProperties.setProperty("javax.persistence.schema-generation.scripts.action", "update");
+			if (settings.getAction() == Action.UPDATE) {
+				outputFile = FileResolver.resolveNextMigrationFile(settings.getMigrationsDir());
+			}
 
-			jpaProperties.setProperty("hibernate.hbm2ddl.delimiter", delimiter);
-			jpaProperties.setProperty("hibernate.format_sql", String.valueOf(formatOutput));
+			settings.getJpaProperties().setProperty("javax.persistence.schema-generation.scripts.create-target", outputFile.getAbsolutePath());
+			settings.getJpaProperties().setProperty("javax.persistence.schema-generation.scripts.drop-target", outputFile.getAbsolutePath());
+
+			settings.getJpaProperties().setProperty("hibernate.hbm2ddl.delimiter", settings.getDelimiter());
+			settings.getJpaProperties().setProperty("hibernate.format_sql", String.valueOf(settings.isFormatOutput()));
 		}
 
 		MetadataSources metadata = new MetadataSources(
 				new StandardServiceRegistryBuilder()
-						.applySettings(jpaProperties)
+						.applySettings(settings.getJpaProperties())
 						.build());
 
-		for (String packageName : packages) {
+		for (String packageName : settings.getPackages()) {
 			listClassNamesInPackage(packageName).forEach(metadata::addAnnotatedClassName);
 		}
 
-		if (generationMode == GenerationMode.METADATA) {
+		if (settings.getGenerationMode() == GenerationMode.METADATA) {
 			SchemaExport export = new SchemaExport();
-			export.setFormat(formatOutput);
-			export.setDelimiter(delimiter);
-			export.setOutputFile(schemaPath.toAbsolutePath().toString());
-			export.execute(EnumSet.of(TargetType.SCRIPT), action.toSchemaExportAction(), metadata.buildMetadata());
+			export.setFormat(settings.isFormatOutput());
+			export.setDelimiter(settings.getDelimiter());
+			export.setOutputFile(outputFile.getAbsolutePath());
+			export.execute(EnumSet.of(TargetType.SCRIPT), settings.getAction().toSchemaExportAction(), metadata.buildMetadata());
 		} else {
+			Connection connection = null;
+			if (settings.getAction() == Action.UPDATE) {
+				if (schemaPath.toFile().exists()) {
+					connection = DriverManager.getConnection(DB_URL, "SA", "");
+					String statement = new String(Files.readAllBytes(schemaPath));
+					connection.prepareStatement(statement).execute();
+				}
+			}
+
 			metadata.buildMetadata().buildSessionFactory().close();
+
+			if (connection != null) {
+				connection.close();
+			}
 		}
 
-		// https://github.com/jOOQ/jOOQ/issues/6707
-		if (schemaPath.toFile().exists()) {
-			byte[] bytes = Files.readAllBytes(schemaPath);
-			if (bytes.length > 0 && bytes[bytes.length - 1] == '\n') {
-				Files.write(schemaPath, Arrays.copyOf(bytes, bytes.length - 1));
-			}
+		if (outputFile.exists()) {
+			List<String> lines = Files.readAllLines(outputFile.toPath())
+					.stream()
+					.map(line -> line.replaceAll("HBM2DDL\\.", ""))
+					.collect(Collectors.toList());
+			Files.write(outputFile.toPath(), lines);
 		}
 	}
